requirements.txt
polars==1.36.1
python-dotenv==1.2.1
requests==2.32.5